pad_project/backend/requirements.txt
fastapi
uvicorn
pydantic[email]
motor
pdfplumber
reportlab
Pillow
python-multipart
